# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Autentication Users",
        "descriptions": "Serviço com foco em validações e gerenciamente de permissões de usuarios",
        "contact": {
            "email": "nivaldeir-silva@hotmail.com"
        }
    },
    "paths": {
        "/auth": {
            "post": {
                "summary": "Autenticar o usuário com email e password",
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "password": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "responses": {
                        "404": {
                            "description": "Not Found",
                            "content": {
                                "application/json": {
                                    "schema": {
                                        "$ref": "#/components/schemas/Error"
                                    }
                                }
                            }
                        },
                        "200": {
                            "description": "OK",
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse200"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/google": {
            "post": {
                "summary": "Autenticação de usuario com google",
                "tags": [
                    "Auth"
                ]
            }
        },
        "/user": {
            "post": {
                "summary": "Cadastro de usuario",
                "descriptions": "",
                "tags": [
                    "User"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            },
                            "examples": {
                                "user": {
                                    "value": {
                                        "email": "test@test.com.br",
                                        "username": "test",
                                        "password": "test"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Já existe informação com esse...",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "put": {
                "descriptions": "Atualizando um usuario pelo ID",
                "summary": "Atualizando um usuario pelo ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuario",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            },
                            "examples": {
                                "user": {
                                    "value": {
                                        "username": "optional",
                                        "password": "optional"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "descriptions": "Desativando um usuario pelo ID",
                "summary": "Desativando um usuario pelo ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuario",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object",
                        "properties": {
                            "email": {
                                "type": "string"
                            },
                            "enabled": {
                                "type": "boolean"
                            },
                            "id": {
                                "type": "string"
                            },
                            "username": {
                                "type": "string"
                            },
                            "password": {
                                "type": "string"
                            },
                            "salt": {
                                "type": "string"
                            }
                        }
                    }
                },
                "Role": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string"
                        }
                    }
                },
                "Permission": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string"
                        }
                    }
                },
                "Auth": {
                    "type": "object",
                    "properties": {
                        "password": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        },
                        "username": {
                            "type": "string"
                        }
                    }
                },
                "ResponseError": {
                    "type": "object",
                    "properties": {
                        "error": {
                            "type": "boolean"
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                }
            },
            "AuthResponse200": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object",
                        "properties": {
                            "token": {
                                "type": "string"
                            },
                            "user": {
                                "type": "object"
                            }
                        }
                    }
                },
                "Role": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string"
                        }
                    }
                },
                "Permission": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string"
                        }
                    }
                },
                "Auth": {
                    "type": "object",
                    "properties": {
                        "password": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        },
                        "username": {
                            "type": "string"
                        }
                    }
                },
                "ResponseError": {
                    "type": "object",
                    "properties": {
                        "error": {
                            "type": "boolean"
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                }
            },
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
            "securitySchemes": {
                "bearerAuth": {
                    "type": "http",
                    "scheme": "bearer",
                    "bearerFormat": "JWT"
                }
            }
        }
    }
}